+++ haulage_mgmt/haulage_logistics/report/haulage_custody_report/haulage_custody_report.js
@@ -1,0 +1,5 @@
+frappe.query_reports["Haulage Custody Report"] = {
+	onload(report) {
+		haulage_mgmt.report_common.default_period_filters(report);
+	},
+};