+++ haulage_mgmt/haulage_logistics/report/haulage_trip_report/haulage_trip_report.js
@@ -1,0 +1,6 @@
+frappe.query_reports["Haulage Trip Report"] = {
+	onload(report) {
+		haulage_mgmt.report_common.default_period_filters(report);
+	},
+	formatter: haulage_mgmt.report_common.net_income_formatter,
+};